feat: Refactor secrets management by consolidating configuration file handling and removing legacy code
